requirements.txt
PyYAML==6.0.1
requests==2.31.0
requests[socks]==2.31.0
